fix: bounds checks, timer parity, price parser, empty-input guard
- JsonLineParser: guard p+31 and p+2 advances against end to prevent
  memmem scanning past mmap region on truncated lines
- main: start timers before producer launch for both mergers so
  throughput numbers are comparable
- JsonLineParser: handle bare numeric price in addition to null/quoted
- NanosParser: cast digit subtraction to uint8_t to avoid signed overflow UB
- HierarchyMerger: return early on empty input to avoid current[0] UB

+++ src/ingestion/NanosParser.hpp
@@ -20,7 +20,7 @@ inline NanoTime parse_iso8601_ns(std::string_view s) noexcept {
     const char* p = s.data();
     auto ri = [p](int off, int n) noexcept -> int64_t {
         int64_t v = 0;
-        for (int i = 0; i < n; i++) v = v * 10 + (p[off + i] - '0');
+        for (int i = 0; i < n; i++) v = v * 10 + static_cast<uint8_t>(p[off + i] - '0');
         return v;
     };
